fix for infinite loop
If the canvas was not displayed on the screen before setting
up the editor control, it would get into an infinite loop of a type
because the canvas width would end up getting set to 0. Now,
the code checks to make sure the canvas dimensions do not get
set to 0 first, ignoring the request otherwise.